+++ oas/src/main/kotlin/io/bkbn/kompendium/oas/security/ApiKeyAuth.kt
@@ -1,20 +1,18 @@
 package io.bkbn.kompendium.oas.security
 
 import kotlinx.serialization.Serializable
-import java.util.Locale
 
 // TODO... is there even an official ktor api auth mechanism??
 
 @Serializable
-@Suppress("UnusedPrivateMember")
-class ApiKeyAuth(val `in`: ApiKeyLocation, val name: String) : SecuritySchema {
+class ApiKeyAuth private constructor(val `in`: String, val name: String) : SecuritySchema {
   val type: String = "apiKey"
 
-  enum class ApiKeyLocation {
-    HEADER,
-    QUERY,
-    COOKIE;
+  constructor(location: ApiKeyLocation, name: String) : this(location.value, name)
 
-    override fun toString(): String = name.lowercase(Locale.getDefault())
+  enum class ApiKeyLocation(val value: String) {
+    HEADER("header"),
+    QUERY("query"),
+    COOKIE("cookie");
   }
 }